requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
boto3==1.33.6
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
botocore==1.33.6
    # via
    #   -r requirements/base.txt
    #   boto3
    #   s3transfer
django==3.2.23
    # via
    #   -r requirements/base.txt
    #   django-crum
    #   edx-i18n-tools
    #   openedx-django-pyfs
    #   openedx-filters
django-crum==0.7.9
    # via -r requirements/base.txt
edx-i18n-tools==1.3.0
    # via -r requirements/dev.in
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.1
    # via
    #   -r requirements/base.txt
    #   openedx-django-pyfs
jmespath==1.0.1
    # via
    #   -r requirements/base.txt
    #   boto3
    #   botocore
lazy==1.6
    # via
    #   -r requirements/base.txt
    #   xblock
lxml==4.9.3
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
    #   xblock
mako==1.3.0
    # via
    #   -r requirements/base.txt
    #   xblock
markupsafe==2.1.3
    # via
    #   -r requirements/base.txt
    #   mako
    #   xblock
openedx-django-pyfs==3.4.0
    # via
    #   -r requirements/base.txt
    #   xblock
openedx-filters==1.6.0
    # via -r requirements/base.txt
path==16.7.1
    # via edx-i18n-tools
polib==1.2.0
    # via edx-i18n-tools
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   botocore
    #   xblock
pytz==2023.3.post1
    # via
    #   -r requirements/base.txt
    #   django
    #   xblock
pyyaml==6.0.1
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
    #   xblock
s3transfer==0.8.2
    # via
    #   -r requirements/base.txt
    #   boto3
simplejson==3.19.2
    # via
    #   -r requirements/base.txt
    #   xblock
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   fs
    #   fs-s3fs
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
typing-extensions==4.8.0
    # via
    #   -r requirements/base.txt
    #   asgiref
urllib3==1.26.18
    # via
    #   -r requirements/base.txt
    #   botocore
web-fragments==2.1.0
    # via
    #   -r requirements/base.txt
    #   xblock
webob==1.8.7
    # via
    #   -r requirements/base.txt
    #   xblock
xblock[django]==1.8.1
    # via
    #   -r requirements/base.txt
    #   xblock

# The following packages are considered to be unsafe in a requirements file:
# setuptools
